+++ src/com/jpexs/decompiler/flash/gui/FolderPreviewPanel.java
@@ -99,7 +99,7 @@ public class FolderPreviewPanel extends JPanel {
 
     public FolderPreviewPanel(final MainPanel mainPanel, List<TreeItem> items) {
         this.items = items;
-        cachedPreviews = Cache.getInstance(false, false, "preview");
+        cachedPreviews = Cache.getInstance(false, false, "preview", true);
 
         addMouseListener(new MouseAdapter() {
             @Override